+++ src/hittable_list.h
@@ -40,7 +40,7 @@ public:
 	{
 		hit_record temp_rec;
 		bool hit_anything = false;
-		double closest_so_far = ray_t.max;
+		float closest_so_far = ray_t.max;
 
 		for (const shared_ptr<hittable>& object : objects) {
 			if (object->hit(r, interval(ray_t.min, closest_so_far), temp_rec)) {
